Conflicts:
	src/components/qt_hmi/qml_model/models/InteractionModel.qml
	src/components/qt_hmi/qml_model/models/VrHelpItemsDefault.qml
	src/components/qt_hmi/qml_model_qt5/models/InteractionModel.qml

+++ src/components/application_manager/src/commands/hmi/on_ui_command_notification.cc
@@ -49,10 +49,10 @@ void OnUICommandNotification::Run() {
   LOG4CXX_INFO(logger_, "OnUICommandNotification::Run");
 
   (*message_)[strings::params][strings::function_id] =
-      mobile_apis::FunctionID::eType::OnCommandID;
+      static_cast<int>(mobile_apis::FunctionID::eType::OnCommandID);
 
   (*message_)[strings::msg_params][strings::trigger_source] =
-      mobile_apis::TriggerSource::TS_MENU;
+      static_cast<int>(mobile_apis::TriggerSource::TS_MENU);
   SendNotificationToMobile(message_);
 }
